+++ vitest.config.js
@@ -25,8 +25,10 @@ const configTest = defineConfig({
     },
   },
   test: {
-    deps: {
-      inline: ['vuetify', '@vcmap/ui'],
+    server: {
+      deps: {
+        inline: ['vuetify', '@vcmap/ui'],
+      },
     },
     environment: 'jsdom',
     setupFiles: ['tests/setup.js'],