+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v1.0.9
+
+- deactivate swipeTool beforeDestroy
+- add info url to swipe window
+
 # v1.0.7
 
 - Fixed testRunner